Inline conditions instead of extracted method
